refactor: consolidate VeniceDeployerValidationTest into VeniceDeployerTest
Merged all 29 tests from VeniceDeployerValidationTest.java into VeniceDeployerTest.java,
removing the now-redundant second file. The merged class is package-private, carries the
@SuppressFBWarnings annotation, and is organized by section using the existing comment
conventions. No tests were dropped; no duplicate test names existed between the two files.
